Remove PremiumPrice from DomainTransfer response (#7)
Premium price field will be deprecated in the API, so we should remove
it from the response. This commit changes the TransferDomain method to
accept a new model called DomainTransferInput that contains only the
needed attributes to make the request and remove the PremiumPrice from
the DomainTransfer model that will be responded from the method.